sync-mailgun: split http utils in their own file
